Refactor multistepflow (#25)
As part of the spike and because we wanted to start refactoring the multistep flow, this is the first step. It introduces two new types of classes:

StepChooser: a replacement for the multistepflow: can be used to select the correct step
SteuerlotseStep: replacement for the Steps. Instead of the MultiStepFlow they are now the main part for any logic that should take place at any given step
Another part we wanted to refactor was to use pydantic for checking the eligibility. This adds that. It also restructures the steps a little bit to make it easier to add more steps.


* Add StepChooser class

* Add refactored SteuerlotseSteps

* Add mock_steps for SteuerlotseSteps

* Replace eligibility flow/steps with Stepchooser and SteuerlotseSteps

* Rename endpoint eligibility2 -> eligibility

* Fix missing argument in call to super._main_handle

* Delete EligiblityResult

* Reformat documentation and comments

* Set prev_step and next_step None at ends

They used to be set to '' at the last or first step of a flow

* Use boolean value for validation check

* Put get_session_data into pre_handle

* Reorder functions in steuerlotseStep

* Reformat tests

* Remove setting the first field

* Copy override_session_data to SteuerlotseStep

* Move information from constructor into into class